requirements.txt
scrapy>=2.7.1
pandas>=1.5.2
python-dotenv>=0.21.0
SQLAlchemy>=2.0.0
psycopg2-binary==2.9.5